use mpn_mulmod_bnm1 instead of mpn_mul_fft in wrap-around functions, since the GMP developers say that mpn_mulmod_bnm1 might be put in the public interface, while mpn_mul_fft will never be put there. See http://gmplib.org/list-archives/gmp-devel/2012-October/002545.html

+++ trunk/mpzspv.c
@@ -307,7 +307,7 @@ ntt_print_vec (const char *msg, const spv_t spv, const spv_size_t l,
   gmp_printf ("%s [%Nd", msg, (mp_ptr) spv, 1);
   for (i = 1; i < l; i++)
     gmp_printf (", %Nd", (mp_ptr) spv + i, 1);
-  printf ("] (mod %llu)\n", p);
+  printf ("] (mod %llu)\n", (long long unsigned int) p);
 }
